Add prototype list library. Needs refactoring.
This commit adds a prototype list library, which will eventually be
used as the underlying data structure for out set implementation.

The list library's functions work as advertised, but they aren't
easy to use. In particular, the 'top' gets lost too easily.

A refactor is planned that will add a global 'top' node that sits
at index [-1] of the list always, This will also act as a nice
stopping point for destroy operations.